toy_project/nukki/code/requirements.txt
flask
rembg
pillow
onnxruntime
opencv-python-headless
numpy